chore: move case sensitivity enum to filter types
