+++ src/Components/IndexScene/LineFilterVariablesScene.tsx
@@ -1,14 +1,13 @@
 import { SceneComponentProps, SceneObjectBase, SceneObjectState } from '@grafana/scenes';
 import React, { ChangeEvent, KeyboardEvent } from 'react';
 import { getLineFiltersVariable } from '../../services/variableGetters';
-import { LineFilterOp } from '../../services/filterTypes';
+import { LineFilterCaseSensitive, LineFilterOp } from '../../services/filterTypes';
 import { reportAppInteraction, USER_EVENTS_ACTIONS, USER_EVENTS_PAGES } from '../../services/analytics';
 import { AdHocFilterWithLabels } from '../../services/scenes';
 import { debounce } from 'lodash';
 import { GrafanaTheme2 } from '@grafana/data';
 import { css } from '@emotion/css';
 import { IconButton, useStyles2 } from '@grafana/ui';
-import { LineFilterCaseSensitive } from '../ServiceScene/LineFilter/LineFilterScene';
 import { LineFilterEditor, LineFilterEditorProps } from '../ServiceScene/LineFilter/LineFilterEditor';
 
 interface LineFilterRendererState extends SceneObjectState {}
